Refactor according to PR review
